+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Removed usages of Random in unit tests to ensure the tests are deterministic.
 * **validate** will now run on all the pack content items when the pack supported marketplaces are modified.
 * Fixed an issue where errors in **validate** were logged as `info`.
+* **lint** will now fail on `demisto.results` and `return_outputs` usage, when pack is `xsoar` or `partner` supported.
+* **lint** will now fail on `LOG` usage in python files.
 
 ## 1.17.2
 * Fixed an issue where **lint** and **validate** commands failed on integrations and scripts that use docker images that are not available in the Docker Hub but exist locally.